feat: replace command runner
The command runner method was using cmd as shell to work with which i couldn't find a way to run commands as administrator, so now i decided to replace it with new code. In new code I use powershell as my commands runner shell which is more powerful and I can config it so much easier. This code runs the income string as adminitrator and some more configs